+++ pymc3/step_methods/__init__.py
@@ -21,13 +21,13 @@ from .metropolis import DEMetropolis, DEMetropolisZ
 from .metropolis import BinaryMetropolis
 from .metropolis import BinaryGibbsMetropolis
 from .metropolis import CategoricalGibbsMetropolis
+from .mlda import MLDA, MetropolisMLDA
 from .metropolis import NormalProposal
 from .metropolis import CauchyProposal
 from .metropolis import LaplaceProposal
 from .metropolis import PoissonProposal
 from .metropolis import MultivariateNormalProposal
-from .metropolis import RecursiveDAProposal
-from .metropolis import MLDA
+from .mlda import RecursiveDAProposal
 
 from .gibbs import ElemwiseCategorical
